Move AfterDisplayMixin to utils/mixin
