+++ lib/view/launch_page.dart
@@ -1,17 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_training/utils/mixin/after_display_mixin.dart';
 import 'package:flutter_training/view/weather_view/weather_page.dart';
-
-mixin AfterDisplayMixin<T extends StatefulWidget> on State<T> {
-  void afterDisplay();
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.endOfFrame.then((_) {
-      afterDisplay();
-    });
-  }
-}
 
 class LaunchPage extends StatefulWidget {
   const LaunchPage({super.key});
